Level design + resolved git issues
Level build. Added kitchen and craftinng room. Crates with no rotation will be used to be moved in the level have made the kitchen more complicated so the player can problem solve in the left corner of the kitchen will be a fire place and that can spit fire. we need to talk more about ideas and what we can both do but this is defo managable. Very sorry for the long wait and the problem solving with github!